Main files for hack
There is one server, one admin and as many clients as you want at one time. Without the server bound on a forwarded port on an IP, this hack is useless. 

SERVER:
The whole program is standstill without this core component. You can run this server on anything. I used to run it on an old jailbroken iPhone 24/7 and it worked very well. Recommended to be used with Raspberry Pi for greater long term reliability but it can work on anything with Python 3.7. 

ADMIN_CLIENT:
This Admin client is the client that is able to control the other clients through the server with certain commands

CLIENT:
This is what is planted on the victims PC and it can be placed in the startup folder with a shortcut of the file once compiled which can be used to recieve commands and follow them accordingly.